requirements.txt
biopython>=1.83
pdbfixer>=1.9
anarci>=2020.04.23
pandas>=2.2.0
openmm>=8.1.1